+++ src/i18next.js
@@ -103,12 +103,14 @@ class I18n extends EventEmitter {
     // TODO: COMPATIBILITY remove this
     if (this.options.compatibilityAPI === 'v1') compat.appendBackwardsAPI(this);
 
-    this.changeLanguage(this.options.lng, (err, t) => {
-      this.emit('initialized', this.options);
-      this.logger.log('initialized', this.options);
+    setTimeout(() => {
+      this.changeLanguage(this.options.lng, (err, t) => {
+        this.emit('initialized', this.options);
+        this.logger.log('initialized', this.options);
 
-      callback(err, t);
-    });
+        callback(err, t);
+      });
+    }, 10);
 
     return this;
   }